issue #6 (refactor - test): refactored integration tests structure
