+++ test/integration_tests/exercises/exercise.integration.test.ts
@@ -2,6 +2,7 @@ import request from "supertest";
 import { app } from "../../../src";
 import { setup, teardown } from "../utils/setup";
 import { IExercise } from "../../../src/models/exercise";
+import { ITopic } from "../../../src/models/topic";
 
 describe("Exercice integration tests suite", () => {
      jest.setTimeout(60 * 1000);
@@ -22,6 +23,9 @@ describe("Exercice integration tests suite", () => {
 
                // Then
                expect(response.statusCode).toBe(200);
+
+               const exercises: IExercise[] = response.body;
+               expect(exercises.length).toBe(0);
           });
 
           test("2 - Create a new exercice", async () => {
